workaround for Duplicate title
